Checked and renamed prelaunch
updated prelaunch name to 00-prelaunch.

added an abs() to line 33 in 00-Prelaunch.cpp. pls check validity of that
